requirements.txt
annotated-types==0.7.0
backoff==2.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
coinbase-advanced-py==1.8.2
cryptography==44.0.0
idna==3.10
numpy==2.0.2
prometheus_client==0.21.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
websockets==13.1
